Complete the mov semantic

+++ src/contextHandler/PINContextHandler.cpp
@@ -143,7 +143,7 @@ static bool isAddressMapped(ADDRINT addr) {
 }
 
 /* Used to deref a pointer address and returns the targeted byte by size of read */
-uint64_t PINContextHandler::getMemoryValue(uint64_t mem, uint64_t readSize) const
+uint64_t PINContextHandler::getMemoryValue(uint64_t mem, uint32_t readSize) const
 {
 
   if (isAddressMapped(mem) == false){
